refactor: destroyNews
